WeaFer Station can now send a json string through a http request.

+++ Inc/I2CSensor.h
@@ -4,8 +4,8 @@
 #include "stm32f0xx_hal.h"
 
 void I2CSensorInit(I2C_HandleTypeDef *);
-int I2CGetTemperature();
-int I2CGetHumidity();
-int I2CGetPressure();
+int16_t I2CGetTemperature();
+uint16_t I2CGetHumidity();
+uint32_t I2CGetPressure();
 
 #endif /* I2CSENSOR_H_ */
